+++ db/migrate/20210511105439_add_current_school_id_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddCurrentSchoolIdToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :early_career_payments_eligibilities, :current_school, type: :uuid, foreign_key: {to_table: :schools}
+  end
+end