+++ spec/factories/early_career_payments/eligibilities.rb
@@ -2,6 +2,7 @@ FactoryBot.define do
   factory :early_career_payments_eligibility, class: "EarlyCareerPayments::Eligibility" do
     trait :eligible do
       nqt_in_academic_year_after_itt { true }
+      current_school { School.find(ActiveRecord::FixtureSet.identify(:penistone_grammar_school, :uuid)) }
       employed_as_supply_teacher { false }
       subject_to_formal_performance_action { false }
       subject_to_disciplinary_action { false }